test: 그룹 생성 권한 체크 도메인 서비스 테스트
